requirements.txt
# Core runtime dependencies for listener.py / otto.py / vision.py.
click
google-genai
mss
Pillow
pydantic
pynput
python-dotenv